Added folder refactor and entities
