--- src/core/Account/domain/AccountId.ts
@@ -1,3 +1,0 @@
-import { Uuid } from '../../Shared/domain/value-object/Uuid';
-
-export class AccountId extends Uuid {}
