--- src/core/Account/domain/AddressNumber.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-export class AddressNumber extends StringValueObject {}
