--- src/core/Account/domain/CompanyName.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-export class CompanyName extends StringValueObject {}
