--- src/core/Account/domain/Country.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-export class Country extends StringValueObject {}
