--- src/core/Account/domain/PostalCode.ts
@@ -1,2 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-export class PostalCode extends StringValueObject {}
