--- src/core/Account/domain/Region.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-export class Region extends StringValueObject {}
