--- src/core/Account/domain/Street.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-export class Street extends StringValueObject {}
