--- src/core/Account/domain/Vat.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-export class Vat extends StringValueObject {}
