+++ src/core/Account/domain/value-object/AccountId.ts
@@ -1,0 +1,3 @@
+import { Uuid } from '../../../Shared/domain/value-object/Uuid';
+
+export class AccountId extends Uuid {}
