+++ src/core/Account/domain/value-object/CompanyName.ts
@@ -1,0 +1,3 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+
+export class CompanyName extends StringValueObject {}
