+++ src/core/Account/domain/value-object/Country.ts
@@ -1,0 +1,3 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+
+export class Country extends StringValueObject {}
