+++ src/core/Account/domain/value-object/PostalCode.ts
@@ -1,0 +1,2 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+export class PostalCode extends StringValueObject {}
