+++ src/core/Account/domain/value-object/Street.ts
@@ -1,0 +1,3 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+
+export class Street extends StringValueObject {}
