+++ src/core/Account/domain/value-object/Vat.ts
@@ -1,0 +1,3 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+
+export class Vat extends StringValueObject {}
