--- src/core/Category/domain/CategoryId.ts
@@ -1,3 +1,0 @@
-import { Uuid } from '../../Shared/domain/value-object/Uuid';
-
-export class CategoryId extends Uuid {}
