+++ src/core/Category/domain/value-object/CategoryId.ts
@@ -1,0 +1,3 @@
+import { Uuid } from '../../../Shared/domain/value-object/Uuid';
+
+export class CategoryId extends Uuid {}
