+++ src/core/Contact/domain/value-object/ContactId.ts
@@ -1,0 +1,3 @@
+import { Uuid } from '../../../Shared/domain/value-object/Uuid';
+
+export class ContactId extends Uuid {}
